requirements.txt
# Core data manipulation and analysis
pandas>=1.5.0
numpy>=1.21.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Data source integrations (optional)
yfinance>=0.2.0  # For Yahoo Finance data
requests>=2.28.0  # For API calls
python-binance>=1.0.17  # For Binance API integration
ccxt>=4.0.0  # Alternative cryptocurrency exchange library

# Statistical analysis
scipy>=1.9.0

# Development and testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0

# Jupyter notebook support (optional)
jupyter>=1.0.0
ipykernel>=6.15.0

# Logging and configuration
python-dateutil>=2.8.0

# Performance optimization
numba>=0.56.0  # Optional: for numerical acceleration